+++ test/datasplash/api_test.clj
@@ -4,12 +4,15 @@
             [clojure.java.io :as io]
             [clojure.string :as str]
             [clojure.test :refer :all]
+            [clojure.tools.logging :as log]
             [datasplash
-             [api :as ds]]
+             [api :as ds]
+             [core :refer [clj->kv]]]
             [clj-time.core :as time]
             [me.raynes.fs :as fs])
   (:import [org.apache.beam.sdk.testing TestPipeline PAssert]
-           [java.io PushbackReader]))
+           [java.io PushbackReader]
+           [java.util.zip GZIPOutputStream GZIPInputStream]))
 
 (defn glob-file
   [path]
@@ -59,13 +62,23 @@
 
 (def test-data [1 2 3 4 5])
 (def json-file-path "json-test-input.json")
+(def gzipped-json-file-path "json-test-input.ndjson.gz")
+
+(defn write-data
+  [path-or-stream data & {:keys [compression]}]
+  (let [compression-fn (case compression
+                         :gzip #(GZIPOutputStream. %)
+                         identity)]
+    (with-open [wrtr (io/writer (compression-fn (io/output-stream path-or-stream)))]
+      (.write wrtr (str/join "\n" (for [l data] (json/encode l)))))))
 
 (defn create-json-input-fixture
   [f]
-  (spit json-file-path (str/join "\n" (for [l test-data]
-                                        (json/encode l))))
+  (do (write-data json-file-path test-data)
+      (write-data gzipped-json-file-path test-data :compression :gzip))
   (f)
-  (fs/delete json-file-path))
+  (do (fs/delete json-file-path)
+      (fs/delete gzipped-json-file-path)))
 
 (use-fixtures :once create-json-input-fixture)
 
@@ -83,7 +96,7 @@
           pipe (ds/->> :pipelined input
                        (ds/map inc {:name :inc :intra-bundle-parallelization 5})
                        (ds/filter even? {:name :even? :intra-bundle-parallelization 5}))
-          output (ds/write-edn-file intra-bundle-parallelization-test {:without-sharding true} pipe)]
+          output (ds/write-edn-file intra-bundle-parallelization-test {:num-shards 1} pipe)]
       (ds/run-pipeline p))
     (let [res (into #{} (read-file (first (glob-file intra-bundle-parallelization-test))))]
       (is (= res #{2 4 6})))))
@@ -95,7 +108,7 @@
           pipe (ds/->> :pipelined input
                        (ds/map inc {:name :inc})
                        (ds/filter even? {:name :even?}))
-          output (ds/write-edn-file pt-test {:without-sharding true} pipe)]
+          output (ds/write-edn-file pt-test {:num-shards 1} pipe)]
       (ds/run-pipeline p))
     (let [res (into #{} (read-file (first (glob-file pt-test))))]
       (is (= res #{2 4 6})))))
@@ -107,7 +120,7 @@
           pipe (ds/cond->> :pipelined input
                            true (ds/map inc {:name :inc})
                            false (ds/filter even? {:name :even?}))
-          output (ds/write-edn-file pt-cond-test {:without-sharding true} pipe)]
+          output (ds/write-edn-file pt-cond-test {:num-shards 1} pipe)]
       (ds/run-pipeline p))
     (let [res (into #{} (read-file (first (glob-file pt-cond-test))))]
       (is (= res #{2 3 4 5 6})))))
@@ -120,7 +133,7 @@
           proc (ds/map (fn [x] (get-in (ds/side-inputs) [:mapping x]))
                        {:side-inputs {:mapping side-input}}
                        input)
-          output (ds/write-edn-file side-test {:without-sharding true} proc)]
+          output (ds/write-edn-file side-test {:num-shards 1} proc)]
       (ds/run-pipeline p))
     (let [res (into #{} (read-file (first (glob-file side-test))))]
       (is (= res #{:a :b :c :d :e})))))
@@ -131,8 +144,8 @@
           input (ds/generate-input [1 2 3 4 5] {:name :main-gen} p)
           {:keys [simple multi]} (ds/map (fn [x] (ds/side-outputs :simple x :multi (* x 10)))
                                          {:side-outputs [:simple :multi]} input)
-          output-simple (ds/write-edn-file sideout-simple-test {:without-sharding true} simple)
-          output-multi (ds/write-edn-file sideout-multi-test {:without-sharding true} multi)]
+          output-simple (ds/write-edn-file sideout-simple-test {:num-shards 1} simple)
+          output-multi (ds/write-edn-file sideout-multi-test {:num-shards 1} multi)]
       (ds/run-pipeline p))
     (let [res-simple (into #{} (read-file (first (glob-file sideout-simple-test))))
           res-multi (into #{} (read-file (first (glob-file sideout-multi-test))))]
@@ -144,7 +157,7 @@
     (let [p (ds/make-pipeline [])
           input (ds/generate-input [{:key :a :val 42} {:key :b :val 56} {:key :a :lue 65}] p)
           grouped (ds/group-by :key {:name "group"} input)
-          output (ds/write-edn-file group-test {:without-sharding true} grouped)]
+          output (ds/write-edn-file group-test {:num-shards 1} grouped)]
       (is "group" (.getName grouped))
       (ds/run-pipeline p)
       (let [res (->> (read-file (first (glob-file group-test)))
@@ -154,21 +167,46 @@
         (is (res [:b #{{:key :b :val 56}}]))))))
 
 (deftest cogroup-test
-  (with-files [cogroup-test]
-    (let [p (ds/make-pipeline [])
-          input1 (ds/generate-input [{:key :a :val 42} {:key :b :val 56} {:key :a :lue 65}] {:name :gen1} p)
-          input2 (ds/generate-input [{:key :a :lav 42} {:key :a :uel 65} {:key :c :foo 42}] {:name :gen2} p)
-          grouped (ds/cogroup-by {:name "cogroup-test"}
-                                 [[input1 :key] [input2 :key]])
-          output (ds/write-edn-file cogroup-test {:without-sharding true} grouped)]
-      (ds/run-pipeline p)
-      (is "cogroup-test" (.getName grouped))
-      (let [res (->> (read-file (first (glob-file cogroup-test)))
-                     (map (fn [[k [i1 i2]]] [k [(into #{} i1) (into #{} i2)]]))
-                     (into #{}))]
-        (is (= res #{[:a [#{{:key :a, :lue 65} {:key :a, :val 42}} #{{:key :a, :uel 65} {:key :a, :lav 42}}]]
-                     [:c [#{} #{{:key :c, :foo 42}}]]
-                     [:b [#{{:key :b, :val 56}} #{}]]}))))))
+  (testing "nominal case"
+    (with-files [cogroup-test]
+      (let [p (ds/make-pipeline [])
+            input1 (ds/generate-input [{:key :a :val 42} {:key :b :val 56} {:key :a :lue 65}] {:name :gen1} p)
+            input2 (ds/generate-input [{:key :a :lav 42} {:key :a :uel 65} {:key :c :foo 42}] {:name :gen2} p)
+            grouped (ds/cogroup-by {:name "cogroup-test"}
+                                   [[input1 :key] [input2 :key]])
+            output (ds/write-edn-file cogroup-test {:num-shards 1} grouped)]
+        (ds/run-pipeline p)
+        (is "cogroup-test" (.getName grouped))
+        (let [res (->> (read-file (first (glob-file cogroup-test)))
+                       (map (fn [[k [i1 i2]]] [k [(into #{} i1) (into #{} i2)]]))
+                       (into #{}))]
+          (is (= res #{[:a [#{{:key :a, :lue 65} {:key :a, :val 42}} #{{:key :a, :uel 65} {:key :a, :lav 42}}]]
+                       [:c [#{} #{{:key :c, :foo 42}}]]
+                       [:b [#{{:key :b, :val 56}} #{}]]}))))))
+
+  (testing "large set of pcollection"
+    (with-files [cogroup-test]
+      (let [p (ds/make-pipeline [])
+            ;; pcolls is something like
+            ;; [ [{:i 0 :key 0}, {:i 1 :key 0} ...] 🠐 this is a pcoll
+            ;;   [{:i 0 :key 1}, {:i 1 :key 1} ...]
+            ;;   ... ]
+            nb-pcolls 101
+            pcolls (mapv (fn [k-pcoll]
+                           (ds/generate-input
+                            (map (fn [i] {:i i :key k-pcoll}) (range 5))
+                            p))
+                         (range nb-pcolls))
+            grouped (ds/cogroup-by {:name "join-fitments"
+                                    :collector (fn [[id_ same-i]] same-i)}
+                                   (mapv #(vector % :i) pcolls))
+            _output (ds/write-edn-file cogroup-test {:num-shards 1} grouped)]
+        (ds/run-pipeline p)
+        (doseq [line (read-file (first (glob-file cogroup-test)))
+                :let [same-i (mapcat identity line)]]
+          (is (= 1 (count (distinct (map :i same-i)))))
+          (is (= (range nb-pcolls) (map :key same-i)) ))
+        ))))
 
 (deftest cogroup-drop-nil-test
   (with-files [cogroup-drop-nil-test]
@@ -177,7 +215,7 @@
           input2 (ds/generate-input [{:key :a :lav 42} {:uel 65} {:key :c :foo 42}] {:name :gen2} p)
           grouped (ds/cogroup-by {:name "cogroup-drop-nil-test"}
                                  [[input1 :key] [input2 :key {:drop-nil? true}]])
-          output (ds/write-edn-file cogroup-drop-nil-test {:without-sharding true} grouped)]
+          output (ds/write-edn-file cogroup-drop-nil-test {:num-shards 1} grouped)]
       (ds/run-pipeline p)
       (is "cogroup-drop-nil-test" (.getName grouped))
       (let [res (->> (read-file (first (glob-file cogroup-drop-nil-test)))
@@ -194,7 +232,7 @@
           input2 (ds/generate-input [{:key :a :lav 42} {:key :a :uel 65} {:key :c :foo 42}] {:name :gen2} p)
           grouped (ds/cogroup-by {:name "cogroup-required-test"}
                                  [[input1 :key {:type :required}] [input2 :key]])
-          output (ds/write-edn-file cogroup-required-test {:without-sharding true} grouped)]
+          output (ds/write-edn-file cogroup-required-test {:num-shards 1} grouped)]
       (ds/run-pipeline p)
       (is "cogroup-required-test" (.getName grouped))
       (let [res (->> (read-file (first (glob-file cogroup-required-test)))
@@ -210,7 +248,7 @@
           input2 (ds/generate-input [{:key :a :lav 42} {:uel 65} {:key :c :foo 42}] {:name :gen2} p)
           grouped (ds/cogroup-by {:name "cogroup-join-nil-test"}
                                  [[input1 :key] [input2 :key]])
-          output (ds/write-edn-file cogroup-join-nil-test {:without-sharding true} grouped)]
+          output (ds/write-edn-file cogroup-join-nil-test {:num-shards 1} grouped)]
       (ds/run-pipeline p)
       (is "cogroup-join-nil-test" (.getName grouped))
       (let [res (->> (read-file (first (glob-file cogroup-join-nil-test)))
@@ -228,7 +266,7 @@
           input2 (ds/generate-input [{:key :a :lav 42} {:key :a :uel 65} {:key :c :foo 42}] {:name :gen2} p)
           grouped (ds/join-by {:name "join-test"}
                               [[input1 :key] [input2 :key]] merge)
-          output (ds/write-edn-file join-test {:without-sharding true} grouped)]
+          output (ds/write-edn-file join-test {:num-shards 1} grouped)]
       (ds/run-pipeline p)
       (is "join-test" (.getName grouped))
       (let [res (into #{} (read-file (first (glob-file join-test))))]
@@ -244,7 +282,7 @@
                               [[input1 :key]
                                [input2 :key {:type :required}]]
                               merge)
-          output (ds/write-edn-file join-test-required {:without-sharding true} grouped)]
+          output (ds/write-edn-file join-test-required {:num-shards 1} grouped)]
       (ds/run-pipeline p)
       (is "join-test-required" (.getName grouped))
       (let [res (into #{} (read-file (first (glob-file join-test-required))))]
@@ -260,18 +298,30 @@
     (ds/run-pipeline p)))
 
 
-
-(deftest combine-pipeline-map
+(defn- test-combine-fn
+  [combine-fn]
   (let [p (ds/make-pipeline [])
         input (ds/generate-input [{:a 1} {:b 2} {:c 3} {:d 4} {:e 5}] p)
-        proc (ds/combine (ds/combine-fn (fn [acc x] (merge acc x))
-                                        identity
-                                        (fn [& accs] (apply merge accs))
-                                        (fn [] {}))
+        proc (ds/combine combine-fn
                          {:name "combine" :scope :global} input)]
     (-> (PAssert/that proc) (.containsInAnyOrder #{{:a 1 :b 2 :c 3 :d 4 :e 5}}))
     (is "combine" (.getName proc))
     (ds/run-pipeline p)))
+
+
+(deftest combine-pipeline-map
+  (let [reducef (fn [acc x] (merge acc x))
+        extractf identity
+        combinef (fn [& accs] (apply merge accs))
+        initf (fn [] {})]
+
+    (test-combine-fn (ds/combine-fn reducef extractf combinef initf))
+
+    (doseq [m [{:reduce reducef :extract extractf :combine combinef :init initf}
+               {:reduce reducef                   :combine combinef :init initf}
+               {:reduce merge                                       :init initf}
+               {:reduce merge}]]
+      (test-combine-fn (ds/combine-fn m)))))
 
 (deftest combine-juxt
   (with-files [combine-juxt-test]
@@ -286,7 +336,7 @@
                             (ds/count-fn) (ds/count-fn :predicate even?)
                             (ds/max-fn :mapper #(* 10 %)))
                            {:name "combine"} input)
-          output (ds/write-edn-file combine-juxt-test {:without-sharding true} proc)]
+          output (ds/write-edn-file combine-juxt-test {:num-shards 1} proc)]
       (is "combine" (.getName proc))
       (ds/run-pipeline p)
       (let [res (into #{} (read-file (first (glob-file combine-juxt-test))))]
@@ -302,8 +352,8 @@
           p4 (ds/combine (ds/sum-fn) {:name :input} input)
           all (ds/concat p1 p2 p3 p4)
           ps (ds/sample 2 input)
-          output1 (ds/write-edn-file math-and-diamond-test {:name :output-all :without-sharding true} all)
-          output2 (ds/write-edn-file sample-test {:name :output-sample :without-sharding true} ps)]
+          output1 (ds/write-edn-file math-and-diamond-test {:name :output-all :num-shards 1} all)
+          output2 (ds/write-edn-file sample-test {:name :output-sample :num-shards 1} ps)]
       (ds/run-pipeline p)
       (let [res (read-file (first (glob-file math-and-diamond-test)))]
         (is (= '(1 3.0 5 15) (sort res))))
@@ -314,6 +364,20 @@
   (let [p (ds/make-pipeline [])
         input (ds/generate-input [1 2 3 4 5] p)
         proc (ds/map (fn [e] (ds/with-timestamp (time/now) e)) input)]
+    (ds/run-pipeline p)))
+
+(deftest stateful-map
+  (let [p (->> (ds/make-pipeline [])
+               (ds/generate-input [(clj->kv [:a 1]) (clj->kv [:a 1])
+                                   (clj->kv [:b 2]) (clj->kv [:b 2])]
+                                  {:coder (ds/make-kv-coder)})
+               (ds/map (fn [[k v]]
+                         (let [state (ds/state)
+                               current (or (.read state) 0)]
+                           (.write state v)
+                           (clj->kv [k (+ v current)])))
+                       {:stateful? true}))]
+    (.. PAssert (that p) (containsInAnyOrder #{(clj->kv [:a 1]) (clj->kv [:a 2]) (clj->kv [:b 2]) (clj->kv [:b 4])}))
     (ds/run-pipeline p)))
 
 (deftest windows
@@ -347,3 +411,19 @@
       (let [cp (flatten (map read-file (glob-file checkpoint-test)))]
         (.. PAssert (that p) (containsInAnyOrder #{2 3 4 5}))
         (is (= '(1 2 3 4) (sort cp)))))))
+
+(deftest compression-out-test
+  (with-files [compression-out-test]
+    (let [p (ds/make-pipeline [])
+          input (ds/generate-input [1 2 3] p)
+          output (ds/write-edn-file compression-out-test {:num-shards 1 :compression-type :gzip} input)]
+      (ds/run-pipeline p)
+      (let [res (->> (read-file (java.util.zip.GZIPInputStream. (io/input-stream (first (glob-file compression-out-test)))))
+                     (into #{}))]
+        (is (= res #{1 2 3}))))))
+
+(deftest compression-in-test
+  (let [p (ds/make-pipeline [])
+        input (ds/read-json-file gzipped-json-file-path {:name :read-json :compression-type :gzip} p)]
+    (-> (PAssert/that input)  (.containsInAnyOrder (map int test-data)))
+    (ds/run-pipeline p)))