+++ CHANGELOG.md
@@ -1,5 +1,49 @@
 # Changes for Datasplash
 
+## v0.6.6 (2019-09-11)
+
+**[compare](https://github.com/ngrunwald/datasplash/compare/v0.6.5...v0.6.6)**
+
+### Fixes
+
+- Fix `cogroup-by` for numbers of pcoll > 10 => Issue #83
+- Fix `:many-files` option in `read-text-file` => Issue #75
+
+## v0.6.5 (2019-08-23)
+
+**[compare](https://github.com/ngrunwald/datasplash/compare/v0.6.4...v0.6.5)**
+
+- Add FileIO Read to read a PCollection of files
+- Add ElasticSearchIO
+- Upgrade to Apache Beam 2.15.0
+- Upgrade dependencies
+
+## v0.6.4 (2019-01-16)
+
+**[compare](https://github.com/ngrunwald/datasplash/compare/v0.6.3...v0.6.4)**
+
+- Upgrade to Apache Beam 2.9.0
+- Upgrade dependencies
+
+## v0.6.3 (2018-11-19)
+
+**[compare](https://github.com/ngrunwald/datasplash/compare/v0.6.2...v0.6.3)**
+
+- Add json-schema in bq write
+- Combine-fn accepts now map as arguments
+- Remove deprecated write-by
+
+
+## v0.6.2 (2018-11-15)
+
+**[compare](https://github.com/ngrunwald/datasplash/compare/v0.6.1...v0.6.2)**
+
+- Upgrade to Apache Beam 2.8.0
+- Support dynamic file writes
+- Support `:deflate` and `:zip` compressions
+- `bq/->time-partitioning` now accepts `:field` and `:require-partition-filter`
+- Improve docs
+
 ## v0.6.1 (2018-03-20)
 
 **[compare](https://github.com/ngrunwald/datasplash/compare/v0.6.0...v0.6.1)**